requirements.txt
python-dotenv==1.0.1
discord-py-interactions==5.14.0
requests==2.32.3
Unidecode==1.3.8
sorcery==0.2.2
ruff==0.9.5